telegram-bot/requirements.txt
absl-py==2.0.0
alabaster==0.7.13
antlr4-python3-runtime==4.9.3
anyio==4.0.0
attrs==23.1.0
Babel==2.12.1
black==23.9.1
boto3==1.28.52
botocore==1.31.52
build==1.0.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
coloredlogs==15.0.1
contourpy==1.1.1
coverage==5.3.1
cycler==0.11.0
Deprecated==1.2.14
docutils==0.17.1
einops==0.3.2
exceptiongroup==1.1.3
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.42.1
future==0.18.3
google-auth==2.23.0
google-auth-oauthlib==1.0.0
grpcio==1.58.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
imagesize==1.4.1
Jinja2==3.1.2
jmespath==1.0.1
json-tricks==3.16.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lit==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.0
onnx-simplifier==0.4.33
onnxruntime==1.13.1
opencv-python==4.8.0.76
packaging==23.1
pathspec==0.11.2
Pillow==10.0.1
pip-tools==7.3.0
platformdirs==3.10.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pyDeprecate==0.3.2
Pygments==2.16.1
pyparsing==2.4.5
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.5
PyYAML==6.0.1
rapidfuzz==3.3.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.3
rpds-py==0.10.3
rsa==4.9
s3transfer==0.6.2
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==4.0.3
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stringcase==1.2.0
super-gradients==3.2.1
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==1.1.0
tomli==2.0.1
torch==2.0.1
torchmetrics==0.8.0
torchvision==0.15.2
tqdm==4.66.1
treelib==1.6.1
triton==2.0.0
typing_extensions==4.8.0
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
